feat: added fields display
AB#75
