+++ webspot/web/static/js/components/nav/top-navbar.js
